Updated Motherboard M2 slot collapsing logic and other fixes

+++ Backend/src/PcBuilderBackend.Application/Catalog/GraphicsCards/Dto/GraphicsCardListItemDto.cs
@@ -17,7 +17,7 @@ public record GraphicsCardListItemDto
     public decimal LengthMm { get; init; }
     public decimal WidthMm { get; init; }
     public decimal HeightMm { get; init; }
-    public decimal PowerConsumptionWatts { get; init; }
+    public int PowerConsumptionWatts { get; init; }
     public PsuCableType PowerConnectorType { get; init; }
     public int PowerConnectorCount { get; init; }
 }
